refactor: 桌宠点击穿透轮询移出 lib.rs 收拢到 api::pet
lib.rs 只保留一行 spawn_hit_test_poll 调用，轮询循环
（pet:cursor 去抖广播 + solid 区域命中测试）整体迁入
api/pet.rs——该模块本就拥有 HitTestState/CursorPosition/Rect
等类型，领域归属一致；逻辑与注释原样搬迁，无行为变化。
依赖 CI 验证编译。
